Refactor PNDA CLI into deployment specific modules
Refactor the PNDA CLI to separate the functionality for AWS and
existing machines into separate modules.

PNDA-4615
